feat(error): add error controller to project
