+++ src/controller/ErrorController.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\src\controller;
+
+class ErrorController
+{
+  public function errorNotFound()
+  {
+    require '../templates/errors/error_404.php';
+  }
+
+  public function errorServer()
+  {
+    require '../templates/errors/error_500.php';
+  }
+}
