+++ src/controller/FrontController.php
@@ -1,33 +1,11 @@
 <?php
 
-namespace App\config;
+namespace App\src\controller;
 
-use App\src\controller\FrontController;
-use Exception;
-
-class Router
+class FrontController
 {
-  private $frontController;
-
-  public function __construct()
+  public function home()
   {
-    $this->frontController = new FrontController();
-  }
-
-  public function run()
-  {
-    try {
-      if (isset($_GET['route'])) {
-        if ($_GET['route'] === 'article') {
-          print_r('pas d\'article pour le moment');
-        } else {
-          print_r('page inconnue');
-        }
-      } else {
-        $this->frontController->home();
-      }
-    } catch (Exception $e) {
-      print_r('Erreur');
-    }
+    require '../templates/home.php';
   }
 }
